% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{appendToTree}
\alias{appendToTree}
\title{Extended Merging Path Plot}
\usage{
appendToTree(factorMerger, plot)
}
\arguments{
\item{factorMerger}{Object of a class \code{factorMerger}}

\item{plot}{Object of a class \code{gg}. There is a variety of plots implemented in
the \code{factorMerger} package highly recommended to use as the \code{plot} parameter.}
}
\description{
Plots Merging Path Plot with additional summary.
}
\examples{
## Multinomial gaussian model
randSample <- generateMultivariateSample(N = 100, k = 10, d = 3)
fm <- mergeFactor(randSample$response, randSample$factor)
appendToTree(fm, plotProfile(fm)) # profile plot on the right
appendToTree(fm, plotGIC(fm)) # GIC plot on the top

}
