% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{BRCA}
\alias{BRCA}
\title{Breast cancer dataset}
\usage{
data(BRCA)
}
\description{
A clinical dataset from The Cancer Genome Atlas Project from
the \code{RTCGA.clinical} package (\url{https://github.com/RTCGA/RTCGA.clinical}).
The variables are as follows:
}
\details{
\itemize{
  \item times event time / censoring time
  \item bcr_patient_barcode patient code
  \item patient.vital_status patient status
  \item patient.drugs.drug.drug_name drug name given to the patient
  \item patient.clinical_cqcf.histological_type patient histological type
  \item patient.stage_event.tnm_categories.pathologic_categories.pathologic_t pathologic type
}
}
\keyword{data}
