% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculateModelStatistic.default}
\alias{calculateModelStatistic.default}
\title{Calculate model statistic (gaussian case) - ...}
\usage{
\method{calculateModelStatistic}{default}(model)
}
\description{
Calculate model statistic (gaussian case) - ...
}
