% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculateModelStatistic.gaussianFactorMerger}
\alias{calculateModelStatistic.gaussianFactorMerger}
\title{Calculate model statistic (gaussian case) - ...}
\usage{
\method{calculateModelStatistic}{gaussianFactorMerger}(factorMerger)
}
\description{
Calculate model statistic (gaussian case) - ...
}
