% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculateModelStatistic.multivariateFactorMerger}
\alias{calculateModelStatistic.multivariateFactorMerger}
\title{Calculate model statistic (multivariate response case) - ...}
\usage{
\method{calculateModelStatistic}{multivariateFactorMerger}(factorMerger)
}
\description{
Calculate model statistic (multivariate response case) - ...
}
