% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculateModelStatistic.nonparametricFactorMerger}
\alias{calculateModelStatistic.nonparametricFactorMerger}
\title{Calculate model statistic (nonparametric case) - ...}
\usage{
\method{calculateModelStatistic}{nonparametricFactorMerger}(factorMerger)
}
\description{
Calculate model statistic (nonparametric case) - ...
}
