% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculatePairStatistic.gaussianFactorMerger}
\alias{calculatePairStatistic.gaussianFactorMerger}
\title{Calculate pair statistic (gaussian case) - ...}
\usage{
\method{calculatePairStatistic}{gaussianFactorMerger}(factorMerger, factor,
  groupA, groupB)
}
\description{
Calculate pair statistic (gaussian case) - ...
}
