% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{calculatePairStatistic.multivariateFactorMerger}
\alias{calculatePairStatistic.multivariateFactorMerger}
\title{Calculate multivariate pair statistic - ...}
\usage{
\method{calculatePairStatistic}{multivariateFactorMerger}(factorMerger, factor,
  groupA, groupB)
}
\description{
Calculate multivariate pair statistic - ...
}
