% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculatePairStatistic.nonparametricFactorMerger}
\alias{calculatePairStatistic.nonparametricFactorMerger}
\title{Calculate nonparametric pair statistic - ...}
\usage{
\method{calculatePairStatistic}{nonparametricFactorMerger}(factorMerger, factor,
  groupA, groupB)
}
\description{
Calculate nonparametric pair statistic - ...
}
