% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculatePairStatistic}
\alias{calculatePairStatistic}
\title{Calculate pair statistic - ...}
\usage{
calculatePairStatistic(factorMerger, factor, groupA, groupB)
}
\description{
Calculate pair statistic - ...
}
