% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerHelper.R
\name{cutree.factorMerger}
\alias{cutree.factorMerger}
\title{Cut a Factor Merger Tree}
\usage{
cutree.factorMerger(factorMerger, stat = "GIC", threshold = NULL)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{stat}{statistic used in a bottom-up search. Available statistics are:
\code{"loglikelihood"}, \code{"pvalue"}, \code{"GIC"}.}

\item{threshold}{cut threshold}
}
\description{
Splits factor levels into non-overlapping clusters based on a \code{factorMerger} object. By default
returns an optimal GIC partition. If a metric and a threshold is specified then performs bottom-up search through models
on the merging path until spots a model scored worse that the given threshold.
}
\details{
By default, \code{cutree} returns factor partition corresponding to the optimal GIC model (with the lowest GIC).
}
