% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerHelper.R
\name{cutTree}
\alias{cutTree}
\title{Cut a Factor Merger Tree}
\usage{
cutTree(factorMerger, stat = "GIC", threshold = NULL)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{stat}{statistic used in the bottom-up search. Available statistics are:
\code{"loglikelihood"}, \code{"pvalue"}, \code{"GIC"}.}

\item{threshold}{cut threshold}
}
\value{
Returns a factor vector - each observation is given a new cluster label.
}
\description{
Splits factor levels into non-overlapping clusters based on a \code{factorMerger} object.
If a metric and a threshold is specified then performs bottom-up search through models
on the merging path until spots a model scored worse than the given threshold.
}
\details{
By default, \code{cutree} returns factor partition corresponding to the optimal GIC model (with the lowest GIC).
}
