% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ess}
\alias{ess}
\title{European Social Survey - happiness}
\usage{
data(ess)
}
\description{
A dataset from \url{http://www.oecd.org/pisa/} about student performance in 2012.
}
\details{
\itemize{
  \item happy
  \item PV1READ student's performance (Reading)
  \item PV1SCIE student's performance (Science)
  \item CNT student's country
  \item OCOD1 parent's occupation
}
}
\keyword{data}
