% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMerger.R, R/factorMergerClasses.R
\docType{package}
\name{factorMerger}
\alias{factorMerger}
\alias{factorMerger-package}
\alias{factorMerger}
\title{factorMerger: ...}
\usage{
factorMerger(response, factor, gaussian = TRUE, subsequent = FALSE)
}
\description{
Description

Factor Merger - ...
}
\section{factorMerger functions}{

}

