% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGeneration.R
\name{generateMultivariateSample}
\alias{generateMultivariateSample}
\title{Generate multivariate sample}
\usage{
generateMultivariateSample(N, k, d = 2, distr = "norm")
}
\description{
Generate multivariate sample
}
