% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{groupsStats}
\alias{groupsStats}
\title{Groups statistic}
\usage{
groupsStats(factorMerger)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}
}
\description{
Summary of statistics specific for a model for each group that appeared in merging.
}
