% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{logLik.mlm}
\alias{logLik.mlm}
\title{Gaussian model loglikelihood}
\usage{
\method{logLik}{mlm}(obj)
}
\description{
http://r.789695.n4.nabble.com/Multiple-regression-information-criterion-td4689113.html
}
