% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{mergeFactors.default}
\alias{mergeFactors.default}
\title{mergeFactors.default}
\usage{
\method{mergeFactors}{default}(x, factor, covariates = NULL, weights = NULL,
  family = "gaussian", method = "fast-adaptive", abbreviate = TRUE)
}
\arguments{
\item{x}{A response \code{vector/matrix} suitable for the model family.}

\item{factor}{A factor \code{vector} when we use \code{response} argument, otherwise the name of column from \code{data} argument containing which levels should be merged.}

\item{covariates}{A covariates \code{vector/matrix}, optional when we use \code{response} argument.}

\item{weights}{A weights \code{vector}, optional when we use \code{response} argument. For more information see: \link[stats]{lm}, \link[stats]{glm}, \link[survival]{coxph}}

\item{family}{Model family to be used in merging. Available models are: \code{"gaussian",}
\code{ "survival", "binomial"}.
By default \code{mergeFactors} uses \code{"gaussian"} model.}

\item{method}{A string specifying method used during merging.}

\item{abbreviate}{Logical. If \code{TRUE}, the default, factor levels names
are abbreviated.}
}
\description{
Default method for \code{mergeFactors()} function.
}
