% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{mergeFactors.formula}
\alias{mergeFactors.formula}
\title{mergeFactors.formula}
\usage{
\method{mergeFactors}{formula}(x, factor, data = NULL, weights = NULL,
  family = "gaussian", method = "fast-adaptive", abbreviate = TRUE)
}
\arguments{
\item{x}{Formula containing columns names from the \code{data} argument.}

\item{factor}{A factor \code{vector} when we use \code{response} argument, otherwise the name of column from \code{data} argument containing which levels should be merged.}

\item{data}{A data frame to be used for modeling}

\item{weights}{A weights \code{vector}, optional when we use \code{response} argument. For more information see: \link[stats]{lm}, \link[stats]{glm}, \link[survival]{coxph}}

\item{family}{Model family to be used in merging. Available models are: \code{"gaussian",}
\code{ "survival", "binomial"}.
By default \code{mergeFactors} uses \code{"gaussian"} model.}

\item{method}{A string specifying method used during merging.}

\item{abbreviate}{Logical. If \code{TRUE}, the default, factor levels names
are abbreviated.}
}
\description{
Method for \code{mergeFactors()} when first argument is a formula.
}
