% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{merger}
\alias{merger}
\title{factorMerger}
\usage{
merger(response, factor, family = "gaussian", successive = FALSE)
}
\description{
\code{factorMerger} is the
base class of the factorMerger package. \code{factorMerger} stores information about
response, initial factor, its levels and their abbreviated names (field \code{map}).
\code{factorMerger} creates its own structure of inheritance connected with model family.
}
