% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{pisa2012}
\alias{pisa2012}
\title{PISA 2012 dataset}
\usage{
data(pisa2012)
}
\description{
A dataset from \url{http://www.oecd.org/pisa/} about student performance in 2012.
}
\details{
\itemize{
  \item PV1MATH student's performance (Mathematics)
  \item PV1READ student's performance (Reading)
  \item PV1SCIE student's performance (Science)
  \item CNT student's country
  \item OCOD1 parent's occupation
}
}
\keyword{data}
