% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot.generatedSample}
\alias{plot.generatedSample}
\title{Generated sample visualization}
\usage{
\method{plot}{generatedSample}(data)
}
\arguments{
\item{x}{generatesSample object}
}
\description{
Generated sample visualization
}
