% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot.generatedSample}
\alias{plot.generatedSample}
\title{Boxplot for generated data}
\usage{
\method{plot}{generatedSample}(data)
}
\arguments{
\item{x}{generatesSample object}
}
\description{
Boxplot for generated data
}

