% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Boxplot (single-dimensional Gaussian)}
\usage{
plotBoxplot(factorMerger, color, clusterSplit)
}
\description{
Plots boxplot with mean as a summary statistic groupping observation by factor levels.
}
