% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotFrequency}
\alias{plotFrequency}
\alias{plotFrequency}
\title{Frequency plot}
\usage{
plotFrequency(factorMerger, colorClusters, clusterSplit)

plotFrequency(factorMerger, colorClusters, clusterSplit)
}
\description{



}
