% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotGIC}
\alias{plotGIC}
\alias{plotGIC}
\title{GIC plot}
\usage{
plotGIC(factorMerger, color, penalty = 2, statistic)

plotGIC(factorMerger, color, penalty = 2, statistic)
}
\description{
Plots Generalized Information Criterion for models on the Factor Merger Tree.

Plots Generalized Information Criterion for models on the Factor Merger Tree.
}
