% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap (multi-dimensional Gaussian)}
\usage{
plotHeatmap(factorMerger, color, clusterSplit)

plotHeatmap(factorMerger, color, clusterSplit)
}
\description{
Plots heatmap for each dimension of the response variable. Vector of means of factor levels for a given
dimension is scaled to have mean equal to zero and standard deviation equal to one.

Plots heatmap for each dimension of the response variable. Vector of means of factor levels for a given
dimension is scaled to have mean equal to zero and standard deviation equal to one.
}
