% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotMeansAndConfInt}
\alias{plotMeansAndConfInt}
\title{Means and standard deviation plot (single-dimensional Gaussian)}
\usage{
plotMeansAndConfInt(factorMerger, color, clusterSplit)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{color}{Boolean. If \code{TRUE}, the default, there is added aesthetic group corresponding
to the final cluster split.}

\item{clusterSplit}{final cluster split definition. A list with two fields:
\itemize{
    \item \code{stat} cluster statistic (available statistics are: \code{"loglikelihood"}, \code{"pvalue"}, \code{"GIC"}),
    \item \code{value} cut threshold / GIC penalty
}}
}
\description{
For each factor level plots its mean and interval of the length equal to its standard deviation.
}
