% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotProfile}
\alias{plotProfile}
\title{Profile plot (multi-dimensional gaussian)}
\usage{
plotProfile(factorMerger)
}
\description{
Plots rank plot - one series is a single factor level and one group
on the OX axis is a single dimension of the response.
}
