% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotProportion}
\alias{plotProportion}
\alias{plotProportion}
\title{Proportion plot (binomial)}
\usage{
plotProportion(factorMerger, color, clusterSplit)

plotProportion(factorMerger, color, clusterSplit)
}
\description{
Plots proportion of success for each factor level.

Plots proportion of success for each factor level.
}
