% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotResponse}
\alias{plotResponse}
\title{Plot Response - Helper Function}
\usage{
plotResponse(factorMerger, responsePanel, colorClusters, clusterSplit,
  responsePanelPalette)
}
\arguments{
\item{factorMerger}{object to be plotted}

\item{responsePanel}{logical - shall it be plotted}

\item{colorClusters}{colors}

\item{clusterSplit}{criteria for splitting}

\item{responsePanelPalette}{colors for reponse}
}
\description{
Plot Response - Helper Function
}
