% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotSurvival}
\alias{plotSurvival}
\alias{plotSurvival}
\title{Survival plot (survival)}
\usage{
plotSurvival(factorMerger, color, clusterSplit)

plotSurvival(factorMerger, color, clusterSplit)
}
\description{
Plots \code{ggcoxadjustedcurves} from the \code{survminer} package.

Plots \code{ggcoxadjustedcurves} from the \code{survminer} package.
}
