% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotTukey}
\alias{plotTukey}
\title{TukeyHSD Plot}
\usage{
plotTukey(factorMerger, palette = NULL)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{palette}{RColorBrewer color palette}
}
\description{
TODO: Aga
}
