% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerHelper.R
\name{predict.factorMerger}
\alias{predict.factorMerger}
\title{Predict method for Factor Merger}
\usage{
\method{predict}{factorMerger}(factorMerger, stat = "GIC", level = NULL)
}
\description{
Predicted values based on \code{factorMerger} object.
}
