% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{print.factorMerger}
\alias{print.factorMerger}
\title{factorMerger}
\usage{
\method{print}{factorMerger}(factorMerger)
}
\description{
\code{factorMerger} is the
base class of the factorMerger package. \code{factorMerger} stores information about
response, initial factor, its levels and their abbreviated names (field \code{map}).
\code{factorMerger} creates its own structure of inheritance connected with model family.

**TODO:** Tell more about the table from print.factorMerger
}
