% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorManipulation.R
\name{setIncreasingOrder}
\alias{setIncreasingOrder}
\title{Set increasing order}
\usage{
setIncreasingOrder(numericVec, factorVec)
}
\arguments{
\item{numericVec}{numeric vector}

\item{factorVec}{factor vector (same length as numbericVec).}
}
\description{
Given a numeric vector and a factor vector changes levels order
in the factor. Means of the numeric variable in groups
appointed by the new order of the factor are increasing.
}

