% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticCalc.R
\name{calculatePairStatistic.multivariateFactorMerger}
\alias{calculatePairStatistic.multivariateFactorMerger}
\title{Calculate pair statistic (multivariate response case) - ...}
\usage{
\method{calculatePairStatistic}{multivariateFactorMerger}(factorMerger, factor,
  groupA, groupB)
}
\description{
Calculate pair statistic (multivariate response case) - ...
}
