% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ess}
\alias{ess}
\title{European Social Survey - happiness}
\usage{
data(ess)
}
\description{
A dataset from \url{http://www.europeansocialsurvey.org/download.html?file=ESS7e02_1&y=2014} on hapiness
of European countries.
}
\details{
\code{ess} is a reweighted version of the original data. The weighting was performed using the \code{dweight} variable with the seed 123.
The variables are as follows:
\itemize{
  \item \code{happy} -- binary, denoting happiness of an invidual. Interviewees were asked
  to assess their happiness in the scale 0-10 (variable \code{happy}). Respondents who gave value
  greater than 5 are considered as "happy",
  \item \code{country} -- respondent's country.
}
}
\keyword{data}
