% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGeneration.R
\name{generateSample}
\alias{generateSample}
\title{Generate sample}
\usage{
generateSample(N, k, distr = "gaussian")
}
\arguments{
\item{N}{sample size}

\item{k}{number of groups}

\item{distr}{type of distribution from \code{c("norm", "exp", "beta")}}
}
\value{
\code{list} with two fields: numeric variable \code{response}
and factor variable \code{factor}.
}
\description{
Produces a random sample of k groups drawn from the same distribution with different
parameters.
}
\examples{
generateSample(100, 2)
generateSample(100, 2, "exp")

}
