% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{groupsStats}
\alias{groupsStats}
\title{Groups statistic}
\usage{
groupsStats(factorMerger)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}
}
\description{
Summary of statistics specific for a model for each group that
appeared in merging.
}
\examples{
randSample <- generateMultivariateSample(N = 100, k = 10, d = 3)
fm <- mergeFactors(randSample$response, randSample$factor)
groupsStats(fm)

}
