% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{mergeFactors}
\alias{mergeFactors}
\title{Merge factors}
\usage{
mergeFactors(response, factor, family = "gaussian",
  method = "fast-adaptive", abbreviate = TRUE)
}
\arguments{
\item{response}{A response \code{vector/matrix} suitable for the model family.}

\item{factor}{A factor \code{vector}.}

\item{family}{Model family to be used in merging. Available models are: \code{"gaussian",}
\code{ "survival", "binomial"}.
By default \code{mergeFactors} uses \code{"gaussian"} model.}

\item{method}{A string specifying method used during merging.
Two methods are availabel: \code{"adaptive", "fast-adaptive", "fixed", "fast-fixed"}.
The default is \code{"fast-adaptive"}.}

\item{abbreviate}{Logical. If \code{TRUE}, the default, factor levels names
are abbreviated.}
}
\description{
Performs step-wise merging of factor levels.
}
\examples{
randSample <- generateMultivariateSample(N = 100, k = 10, d = 3)
mergeFactors(randSample$response, randSample$factor)

}
