% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot.factorMerger}
\alias{plot.factorMerger}
\title{Plot Factor Merger}
\usage{
\method{plot}{factorMerger}(factorMerger, panel = "all",
  show = "loglikelihood", nodesSpacing = "equidistant", summary = NULL,
  color = "null", showDiagnostics = TRUE, alpha = 0.05)
}
\arguments{
\item{panel}{\code{c("all", "response", "GIC", "merging")}}

\item{show}{\code{c("loglikelihood", "p-value")}}

\item{nodesSpacing}{\code{c("equidistant", "effects", "modelSpecific")} # TODO: Change names}

\item{color}{\code{c("null", "cluster", "summary")}}

\item{showDiagnostics}{Boolean}
}
\description{
Plot Factor Merger
}
