% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot.factorMerger}
\alias{plot.factorMerger}
\title{Plot Factor Merger}
\usage{
\method{plot}{factorMerger}(factorMerger, panel = "all",
  show = "loglikelihood", nodesSpacing = "equidistant", summary = NULL,
  color = "none", clusterSplit = list(stat = "GIC", value = 2),
  markBestModel = TRUE, markStars = TRUE, alpha = 0.05, palette = NULL)
}
\arguments{
\item{panel}{\code{c("all", "response", "GIC", "merging")}}

\item{show}{\code{c("loglikelihood", "p-value")}}

\item{nodesSpacing}{\code{c("equidistant", "effects", "modelSpecific")} # TODO: Change names}

\item{color}{\code{c("none", "cluster", "summary")}}

\item{alpha}{significance interval}

\item{palette}{color palette}

\item{showDiagnostics}{Boolean}

\item{penalty}{GIC penalty}
}
\description{
Plot Factor Merger
}
