% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotGIC}
\alias{plotGIC}
\title{GIC plot}
\usage{
plotGIC(factorMerger, color, penalty = 2, statistic)
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{color}{Boolean. If \code{TRUE}, the default, there is added aesthetic group corresponding
to the final cluster split.}

\item{penalty}{GIC penalty}

\item{statistic}{cluster split statistic}
}
\description{
Plots Generalized Information Criterion for models on the Factor Merger Tree.
}
