% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap (multi-dimensional Gaussian)}
\usage{
plotHeatmap(factorMerger, color, clusterSplit, palette = "Greys")
}
\arguments{
\item{factorMerger}{object of a class \code{factorMerger}}

\item{color}{Boolean. If \code{TRUE}, the default, there is added aesthetic group corresponding
to the final cluster split.}

\item{clusterSplit}{final cluster split definition. A list with two fields:
\itemize{
    \item \code{stat} cluster statistic (available statistics are: \code{"loglikelihood"}, \code{"pvalue"}, \code{"GIC"}),
    \item \code{value} cut threshold / GIC penalty
}}

\item{palette}{custom palette}
}
\description{
Plots heatmap for each dimension of the response variable. Vector of means of factor levels for a given
dimension is scaled to have mean equal to zero and standard deviation equal to one.
}
