% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R, R/visualizations.spin.R
\name{plotProfile}
\alias{plotProfile}
\alias{plotProfile}
\title{Profile plot (multi-dimensional Gaussian)}
\usage{
plotProfile(factorMerger, color, clusterSplit)

plotProfile(factorMerger, color, clusterSplit)
}
\description{
Plots rank plot - one series is a single factor level and one group
on the OX axis is a single dimension of the response.

Plots rank plot - one series is a single factor level and one group
on the OX axis is a single dimension of the response.
}
