% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotTree}
\alias{plotTree}
\title{Plot Tree - Helper Function}
\usage{
plotTree(factorMerger, statistic, nodesSpacing, clusterSplit,
  markBestModel, markStars, alpha, color, colorsDf, palette = NULL,
  title, subtitle, panelGrid)
}
\arguments{
\item{factorMerger}{object to be plotted}

\item{statistic}{this parameter will be passed to plotSimpleTree}

\item{nodesSpacing}{this parameter will be passed to plotSimpleTree}

\item{clusterSplit}{this parameter will be passed to plotSimpleTree}

\item{markBestModel}{this parameter will be passed to plotSimpleTree}

\item{markStars}{this parameter will be passed to plotSimpleTree}

\item{alpha}{this parameter will be passed to plotSimpleTree}

\item{color}{this parameter will be passed to plotSimpleTree}

\item{colorsDf}{this parameter will be passed to plotSimpleTree}

\item{palette}{this parameter will be passed to plotSimpleTree}

\item{title}{this parameter will be passed to plotSimpleTree}

\item{subtitle}{this parameter will be passed to plotSimpleTree}

\item{panelGrid}{this parameter will be passed to plotSimpleTree}
}
\description{
Plot Tree - Helper Function
}
