% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMergerClasses.R
\name{print.factorMerger}
\alias{print.factorMerger}
\title{factorMerger}
\usage{
\method{print}{factorMerger}(x, ...)
}
\arguments{
\item{x}{object of a class \code{factorMerger}.}

\item{...}{Other arguments}
}
\description{
\code{factorMerger} is the
base class of the factorMerger package. \code{factorMerger} stores information about
response, initial factor, its levels and their abbreviated names (field \code{map}).
\code{factorMerger} creates its own structure of inheritance connected with model family.

When merging is applied, \code{factorMerger} shows which levels have been
merged together with
the matching summary statistics:
model loglikelihood, pvalue for the \code{LRT} test
against the full model and Generalized Information Criterion value.
}
